requirements.txt
streamlit
pandas
sqlalchemy
streamlit-authenticator
pyyaml
bcrypt
plotly